Update background.png
I don't know why this is necessary or what the diff is, but it failed to
load this image with an error to the effect of "Invalid file format". I
tried opening the png in preview and resaving it but it failed because
background.psd already existed which makes no sense. I deleted the psd
and resaved the image which resulted in this diff.